Bulk checkin delete UI fixes dark mode

+++ resources/views/users/view.blade.php
@@ -276,26 +276,32 @@
             @can('update', $user)
                   @if ($user->deleted_at=='')
                     <div class="col-md-12" style="padding-top: 30px;">
-                        @if ($user->isDeletable())
+                        @if (($user->isDeletable()) && ($user->id!=auth()->user()->id))
                             <a href="" class="delete-asset btn-block btn btn-sm btn-danger btn-social hidden-print" data-toggle="modal" data-title="{{ trans('general.delete') }}" data-content="{{ trans('general.sure_to_delete_var', ['item' => $user->display_name]) }}" data-icon="fa-trash" data-target="#dataConfirmModal" onClick="return false;" >
                                 <x-icon type="delete" />
                                 {{ trans('button.delete')}}
                             </a>
-                            @else
-                            <button class="btn-block btn btn-sm btn-danger btn-social hidden-print disabled">
+                        @elseif ($user->id == auth()->user()->id)
+                            <button class="btn-block btn btn-sm btn-danger btn-social hidden-print disabled" data-tooltip="true" data-title="{{ trans('tooltips.disabled_assoc.user_self') }}">
+                                <x-icon type="delete" />
+                                {{ trans('button.delete')}}
+                            </button>
+                        @else
+                            <button class="btn-block btn btn-sm btn-danger btn-social hidden-print disabled" data-tooltip="true" data-title="{{ trans('tooltips.disabled_assoc.user') }}">
                                 <x-icon type="delete" />
                                 {{ trans('button.delete')}}
                             </button>
                         @endif
                     </div>
                     <div class="col-md-12" style="padding-top: 5px;">
+
                       <form action="{{ route('users/bulkedit') }}" method="POST">
                         <!-- CSRF Token -->
                         <input type="hidden" name="_token" value="{{ csrf_token() }}" />
                         <input type="hidden" name="bulk_actions" value="delete" />
 
                         <input type="hidden" name="ids[{{ $user->id }}]" value="{{ $user->id }}" />
-                        <button class="btn btn-block btn-sm btn-danger btn-social hidden-print">
+                        <button class="btn btn-block btn-sm btn-danger btn-social hidden-print" data-tooltip="true" data-title="{{ trans('tooltips.checkin_all.user') }}">
                             <x-icon type="checkin-and-delete" />
                             {{ trans('button.checkin_and_delete') }}
                         </button>
